Collapse the *_vm.go files into their bases
