+++ appengine.go
@@ -133,3 +133,10 @@ func WithAPICallFunc(ctx context.Context, f APICallFunc) context.Context {
 func APICall(ctx context.Context, service, method string, in, out proto.Message) error {
 	return internal.Call(ctx, service, method, in, out)
 }
+
+// BackgroundContext returns a context not associated with a request.
+// This should only be used when not servicing a request.
+// This only works in App Engine "flexible environment".
+func BackgroundContext() context.Context {
+	return internal.BackgroundContext()
+}
